requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
requests>=2.32.0
jinja2>=3.1.3
weasyprint>=61.0